+++ src/components/Settings/Navigation.vue
@@ -1,10 +1,24 @@
 <template>
     <nav class="settings-nav">
         <RouterLink :to="{name: 'settings.profile'}">Profile</RouterLink>
-        <RouterLink :to="{name: 'settings.categories'}">Ingredient categories</RouterLink>
-        <RouterLink :to="{name: 'settings.glasses'}">Glass types</RouterLink>
+        <template v-if="isAdmin">
+            <RouterLink :to="{name: 'settings.users'}">Users</RouterLink>
+            <RouterLink :to="{name: 'settings.categories'}">Ingredient categories</RouterLink>
+            <RouterLink :to="{name: 'settings.glasses'}">Glass types</RouterLink>
+        </template>
     </nav>
 </template>
+<script>
+import Auth from "@/Auth";
+
+export default {
+    data() {
+        return {
+            isAdmin: Auth.isAdmin()
+        };
+    }
+}
+</script>
 <style scoped>
 .settings-nav {
     display: flex;