+++ src/router/index.js
@@ -95,6 +95,27 @@ const router = createRouter({
                     path: '/settings/glasses/form',
                     name: 'settings.glasses.form',
                     component: () => import('../views/GlassFormView.vue'),
+                },
+                {
+                    path: '/settings/users',
+                    name: 'settings.users',
+                    component: () => import('../views/SettingsUsersView.vue'),
+                },
+                {
+                    path: '/settings/users/form',
+                    name: 'settings.users.form',
+                    component: () => import('../views/UserFormView.vue'),
+                },
+            ]
+        },
+        {
+            path: '/print',
+            component: () => import('../PrintLayout.vue'),
+            children: [
+                {
+                    path: 'cocktail/:id',
+                    name: 'print.cocktail',
+                    component: () => import('../views/CocktailPrintView.vue'),
                 }
             ]
         }