+++ src/views/SettingsUsersView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import UsersList from '../components/Settings/UsersList.vue'
+</script>
+
+<template>
+  <main>
+    <UsersList />
+  </main>
+</template>