+++ src/views/UserFormView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import UserForm from '../components/Settings/UserForm.vue'
+</script>
+
+<template>
+  <main>
+    <UserForm />
+  </main>
+</template>